Add setup and change project structure. Maybe needs further folder refactor. All fixed and tested, now works fine. Known issue: incorrect score_fn leads to uncaught exception.
