+++ .gitignore
@@ -1,3 +1,6 @@
+server/build
+server/dist
+server/geoscore.egg-info
 .idea/
 .vscode/
 __pycache__
