+++ server/Dockerfile
@@ -7,8 +7,10 @@ RUN pip install flask-cors
 
 ## App
 
-ADD . /f1
-WORKDIR /f1
+ADD . /f1-tool
+WORKDIR /f1-tool
+RUN python setup.py install
+WORKDIR /f1-tool/geoscore
 
 CMD ["python3","-u","app.py"]
