+++ server/geoscore/metrics/__init__.py
@@ -1,5 +1,5 @@
 from .objectwise_point_score import objectwise_point_score
 from .areawise_score import areawise_score
-from.objectwise_score import objectwise_score
+from .objectwise_score import objectwise_score
 
 from .common import get_scoring_function
