+++ server/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name='geoscore', version='2.0a', packages=find_packages())
